Decompose VM into composed state modules
